+++ src/commands/tasks/addtasks.js
@@ -8,10 +8,10 @@ module.exports = class extends Command {
       description: 'Rajoute une tâche',
       extendedHelp: 'Aucune aide complémentaire disponible.',
       cooldown: 5,
-      permissionLevel: 0,
+      permissionLevel: 1,
       aliases: ['addtasks', 'addtask', 'add', 'at'],
       runIn: ['text'],
-      usage: '<titre:string> <due_date:date> [description:string]  [...]',
+      usage: '<titre:string> <due_date:date> [description:string] [...]',
       usageDelim: ' '
     })
     this.usageCustom = '%addtask <titre:string> <dueDate:string> [description:string]'
@@ -19,14 +19,7 @@ module.exports = class extends Command {
   }
 
   async run (message, [titre, dueDate, ...description]) {
-    // if the 'addtask' role is set:
-    // adding tasks is restricted to people who have this role
-    const role = message.guild.roles.find(role => role.id === message.guild.settings.roles.addtask)
-    if (role) {
-      if (!message.member.roles.find(r => r.id === role.id)) {
-        return message.reply(`Vous n'avez pas la permission d'utiliser cette commande.`)
-      }
-    }
+    if (!dueDate) return
 
     // sets the task
     // notify user that it has been saved